+++ crates/mako/test/compile/css-modules/user.css
@@ -1,0 +1,4 @@
+
+.user {
+    color: red;
+}
